utils/pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.0"
description = ""
authors = ["KaiErikNiermann <k.e.niermann@student.vu.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
